// frontend/app/controller/model/workflow/workflowDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DocumentDto } from "../Document.model";
import { StepDTO } from "./stepDTO";
import { WorkflowPresetDTO } from "./workflowPresetDTO";

export interface WorkflowDTO { 
    id?: number;
    title?: string;
    description?: string;
    status?: WorkflowDTO.StatusEnum;
    flag?: WorkflowDTO.FlagEnum;
    stepDTOList?:Array<StepDTO>
    dateC?: string;
    dateUpdate?: string;  
    initiateurId: number;
    initiateurNom  ?: string;
    workflowPresetDTO :WorkflowPresetDTO;
    documents?: Array<DocumentDto>;
    piecesJointes?: Array<DocumentDto>;

}
export namespace WorkflowDTO {
    export type StatusEnum = 'OPEN' | 'CLOSED' |'REJECTED'| 'Annulled' | 'REOPENED';
    export const StatusEnum = {
        OPEN: 'OPEN' as StatusEnum,
        CLOSED: 'CLOSED' as StatusEnum,
        REJECTED: 'REJECTED' as StatusEnum,
        Annulled: 'Annulled' as StatusEnum,
        REOPENED: 'REOPENED' as StatusEnum

    };
    export type FlagEnum = 'URGENT' | 'NORMALE' | 'BASSE';
    export const FlagEnum = {
        URGENT: 'URGENT' as FlagEnum,
        NORMALE: 'NORMALE' as FlagEnum,
        BASSE: 'BASSE' as FlagEnum
    };
    export const workflowFlagOptions = [
        { label: 'Haute', tLabel: "priorite.Haute", value: FlagEnum.URGENT },
        { label: 'Moyenne',tLabel: "priorite.Moyenne ", value: FlagEnum.NORMALE },
        { label: 'Basse', tLabel: "priorite.Basse ",value: FlagEnum.BASSE },
    ];

    export const getWorkflowFlagLabel = (flag: FlagEnum | undefined) => {
        return workflowFlagOptions.find((option) => option.value === flag)?.label || '';
    };
}